requirements.txt
google-cloud==0.34.0
google-cloud-aiplatform==1.34.0
vertexai==0.0.1
firebase-admin==6.2.0
langchain==0.0.309
google-cloud-documentai==2.20.0
streamlit==1.27.2
protobuf~=3.20.0